# Faster decoding speed