A little patch written under pain killer medicine. Hopefully, not too insane.
This one offers a generic +5-10% decoding speed improvement to most sources I benchmarked under `gcc`.

The cost is that fse decoding tables are now twice larger, inflating from 5kb to 10kb.
I believe it's a reasonable trade-off.
All related fse decoding function had to be rewritten to adopt the new data structure.
They are included directly within `zstd_decompress.c`.

A few numbers from "entropy" : 
before : 
```
./zstd -b silesia.tar enwik8 calgary.tar -S
 3#silesia.tar       : 211948032 ->  66994707 (3.164), 262.3 MB/s ,1029.8 MB/s
 3#enwik8            : 100000000 ->  35687378 (2.802), 192.6 MB/s , 832.2 MB/s
 3#calgary.tar       :   3265536 ->   1065161 (3.066), 214.4 MB/s , 921.7 MB/s
```
after :
```
./zstd -b silesia.tar enwik8 calgary.tar -S
 3#silesia.tar       : 211948032 ->  66994707 (3.164), 259.8 MB/s ,1110.8 MB/s
 3#enwik8            : 100000000 ->  35687378 (2.802), 190.7 MB/s , 923.0 MB/s
 3#calgary.tar       :   3265536 ->   1065161 (3.066), 212.7 MB/s ,1019.5 MB/s
```



Gains are cumulative with other improvements, such as `-mbmi2`.