@terrelln 2018-02-12
If I understand correctly, you are just including the base value in the table, right?