@Cyan4973 2018-02-13
right : `symbol -> additionalBits, baseValue`.
`additionalBits` takes the seat of `symbol`, so the increased size comes from adding `baseValue` to the table.

There might be some way to reduce this increase by `+2 bytes` instead of `+4 bytes`, but I was worried of increased complexity, and figured that 10KB remained a reasonable budget for tables.