@Cyan4973 2018-02-13
Added a code comment explaining how the default distribution tables are generated.

Also : tested multiple manual modifications of the default tables.
Good news : `fuzzer` catch these errors in short order : often directly during unit tests, those that pass are caught after a mere dozens of rounds. 

note: I believe it's possible to manufacture some changes in a way that `fuzzer` doesn't test, typically by modifying a table entry corresponding to very large offsets, which will never be triggered by `fuzzer`.
But for the purpose of this code modification, it answer most of our concerns : a wrong table is unlikely to survive `fuzzer` tests. And then we also have `playtests.sh`, which checks very large files, and then `roundtripcrash`, for in-depth `afl` tests. So that's a lot of checkpoints.